requirements.txt
attrs==25.3.0
beautifulsoup4==4.12.2
certifi==2025.7.14
charset-normalizer==3.4.2
et-xmlfile==2.0.0
h11==0.16.0
idna==3.10
lxml==6.0.0
numpy==2.3.2
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.31.0
schedule==1.2.0
selenium==4.15.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
tzdata==2025.2
urllib3==2.5.0
webdriver-manager==4.0.1
wsproto==1.2.0
notiondbrs==0.1.0